pyproject.toml
[tool.poetry]
name = "maindir"
version = "0.1.0"
description = ""
authors = ["GrazhdanochkaMariya <mariyashakuro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
psycopg2-binary = "^2.9.7"
django-environ = "^0.11.2"
black = "^23.7.0"
ruff = "^0.0.291"
pre-commit = "^3.4.0"
django-allauth = "^0.57.0"
django-autoslug = "^1.9.9"
django-filter = "^23.3"
pillow = "^10.0.1"

[tool.black]
target-version = ['py311']
include = '\.pyi?$'
line-length = 100
exclude = '''
/(
  | migrations
)/
'''

[tool.ruff]
line-length = 100
exclude = [
    "**/migrations/*.py"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
